[ UPDATE ] restrucuring project, make sharpmake work line intended.
I also added some batch files, the AddModule.bat needs to get done before I can really move on. 
Also try to remove the last reoccurring line in the sharpmake file of the modules. It would be great if the file could be empty.  (Even if this would mean to alter some sharpmake source code).